feat: added tests for escrowpaymentcondition
- fixed CI
