+++ .github/workflows/build.yml
@@ -17,7 +17,7 @@ jobs:
       - uses: nevermined-io/nvm-tools-actions@v0.9.0
         with:
           token: ${{ secrets.API_TOKEN_GITHUB }}
-          node-version: 'develop'
+          contracts-version: "v3.0.0"
 
       - name: Install dependencies
         run: yarn
